requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
botocore==1.34.128
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
exceptiongroup==1.2.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
flask-setup==0.5.5
flask-sock==0.7.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
langchain==0.0.292
langsmith==0.0.92
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.10.0
numpy==1.26.4
openai==0.27.2
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.6.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pyngrok==7.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
tenacity==8.4.1
tqdm==4.66.4
twilio==9.2.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.9.4
